move ors-error to separate file
